Formatting, retryContextError moved to util
